+++ toolchain/scripts/stage1/install_mpich.sh
@@ -47,6 +47,9 @@ case "${with_mpich}" in
         #download_pkg_from_ABACUS_org "${mpich_sha256}" "${mpich_pkg}"
         download_pkg_from_url "${mpich_sha256}" "${mpich_pkg}" "${url}"
       fi
+    if [ "${PACK_RUN}" = "__TRUE__" ]; then
+      echo "--pack-run mode specified, skip installation"
+    else
       echo "Installing from scratch into ${pkg_install_dir} for MPICH device ${MPICH_DEVICE}"
       [ -d mpich-${mpich_ver} ] && rm -rf mpich-${mpich_ver}
       tar -xzf ${mpich_pkg}
@@ -72,6 +75,10 @@ case "${with_mpich}" in
       cd ..
       write_checksums "${install_lock_file}" "${SCRIPT_DIR}/stage1/$(basename ${SCRIPT_NAME})"
     fi
+    fi
+    if [ "${PACK_RUN}" = "__TRUE__" ]; then
+        echo "--pack-run mode specified, skip system check"
+    else
     check_dir "${pkg_install_dir}/bin"
     check_dir "${pkg_install_dir}/lib"
     check_dir "${pkg_install_dir}/include"
@@ -83,6 +90,7 @@ case "${with_mpich}" in
     MPIF77="${MPIFC}"
     MPICH_CFLAGS="-I'${pkg_install_dir}/include'"
     MPICH_LDFLAGS="-L'${pkg_install_dir}/lib' -Wl,-rpath,'${pkg_install_dir}/lib'"
+    fi
     ;;
   __SYSTEM__)
     echo "==================== Finding MPICH from system paths ===================="
