+++ toolchain/scripts/stage3/install_fftw.sh
@@ -42,6 +42,9 @@ case "$with_fftw" in
       else
         download_pkg_from_ABACUS_org "${fftw_sha256}" "${fftw_pkg}"
       fi
+    if [ "${PACK_RUN}" = "__TRUE__" ]; then
+      echo "--pack-run mode specified, skip installation"
+    else
       echo "Installing from scratch into ${pkg_install_dir}"
       [ -d fftw-${fftw_ver} ] && rm -rf fftw-${fftw_ver}
       tar -xzf ${fftw_pkg}
@@ -74,6 +77,7 @@ case "$with_fftw" in
       make install > install.log 2>&1 || tail -n ${LOG_LINES} install.log
       cd ..
       write_checksums "${install_lock_file}" "${SCRIPT_DIR}/stage3/$(basename ${SCRIPT_NAME})"
+    fi
     fi
     FFTW_CFLAGS="-I'${pkg_install_dir}/include'"
     FFTW_LDFLAGS="-L'${pkg_install_dir}/lib' -Wl,-rpath,'${pkg_install_dir}/lib'"
