+++ toolchain/scripts/stage4/install_libri.sh
@@ -43,20 +43,18 @@ case "$with_libri" in
         # download from github.com and checksum
             echo "===> Notice: This version of LibRI is downloaded in GitHub Release, which will always be out-of-date version <==="
             download_pkg_from_url "${libri_sha256}" "${filename}" "${url}"
-            # echo "wget --quiet $url -O $filename"
-            # if ! wget --quiet $url -O $filename; then
-            # report_error "failed to download $url"
-            # recommend_offline_installation $filename $url
-            # fi
-            # # checksum
-            # checksum "$filename" "$libri_sha256"
         fi
+    if [ "${PACK_RUN}" = "__TRUE__" ]; then
+      echo "--pack-run mode specified, skip installation"
+    else
         echo "Installing from scratch into ${pkg_install_dir}"
         [ -d $dirname ] && rm -rf $dirname
         tar -xzf $filename
         cp -r $dirname "${pkg_install_dir}/"
         write_checksums "${install_lock_file}" "${SCRIPT_DIR}/stage4/$(basename ${SCRIPT_NAME})"
     fi
+    fi
+    LIBRI_CFLAGS="-I'${pkg_install_dir}'"
         ;;
     __SYSTEM__)
         echo "==================== CANNOT Finding LIBRI from system paths NOW ===================="
